+++ react/components/PickupSelection.tsx
@@ -1,9 +1,11 @@
 import React from 'react'
 import { useDrawer } from 'vtex.store-drawer/DrawerContext'
+import { useIntl } from 'react-intl'
 
 import PostalCodeInput from './PostalCodeInput'
 import SubmitButton from './SubmitButton'
 import PickupItem from './PickupItem'
+import messages from '../messages'
 
 interface Props {
   onSubmit: (zipCode?: string) => void
@@ -33,16 +35,21 @@ const PickupSelection = ({
   const newZipCodeTyped = zipCode !== selectedZipCode
   const shouldHideUpdateButton = (!zipCode || !newZipCodeTyped) && !isLoading
   const { close } = useDrawer()
+  const intl = useIntl()
 
   return (
     <div className="flex flex-column justify-between">
-      <PostalCodeInput
-        zipCode={zipCode}
-        onSubmit={onSubmit}
-        errorMessage={inputErrorMessage}
-        onChange={onChange}
-        addressLabel={addressLabel}
-      />
+      <div className="mb7">
+        <PostalCodeInput
+          zipCode={zipCode}
+          onSubmit={onSubmit}
+          errorMessage={inputErrorMessage}
+          onChange={onChange}
+          addressLabel={addressLabel}
+          placeholder={intl.formatMessage(messages.postalCodeInputPlaceHolder)}
+          newZipCodeTyped={newZipCodeTyped}
+        />
+      </div>
       <div className="m-100 flex flex-column justify-center">
         {pickups.map((currentPickup) => (
           <PickupItem