+++ react/components/ShippingOptionButton.tsx
@@ -1,10 +1,16 @@
-import React from 'react'
+import React, { useState } from 'react'
 import { useCssHandles } from 'vtex.css-handles'
 import { Spinner } from 'vtex.styleguide'
-
 import '../styles.css'
 
-const CSS_HANDLES = ['buttonWrapper', 'buttonLabel', 'buttonValue'] as const
+import DeliveryPopover from './DeliveryPopover'
+
+const CSS_HANDLES = [
+  'buttonWrapper',
+  'buttonLabel',
+  'buttonValue',
+  'shippingButtonContainer',
+] as const
 
 interface Props {
   onClick: () => void
@@ -13,6 +19,11 @@ interface Props {
   placeholder: string
   value?: string
   compact: boolean
+  zipCode?: string
+  onChange?: (zipCode?: string) => void
+  onSubmit?: () => void
+  inputErrorMessage?: string
+  overlayType?: OverlayType
 }
 
 const ShippingOptionButton = ({
@@ -22,11 +33,33 @@ const ShippingOptionButton = ({
   value,
   placeholder,
   compact,
+  zipCode,
+  onChange,
+  onSubmit,
+  inputErrorMessage,
+  overlayType,
 }: Props) => {
+  const [isPopoverOpen, setIsPopoverOpen] = useState(true)
   const handles = useCssHandles(CSS_HANDLES)
 
+  const handleOutSideClick = () => {
+    setIsPopoverOpen(false)
+  }
+
+  const handlePopoverClick = () => {
+    onClick()
+    setIsPopoverOpen(false)
+  }
+
+  const showPopover =
+    overlayType === 'popover-button' || overlayType === 'popover-input'
+
+  const isFirstLoading = !zipCode && loading
+
+  const openPopover = !isFirstLoading && !value && showPopover && isPopoverOpen
+
   return (
-    <div className="flex items-center">
+    <div className={`${handles.shippingButtonContainer} flex items-center`}>
       <button
         onClick={onClick}
         className={`${handles.buttonWrapper} flex h2 items-center br3 pt3 pb3 b--none`}
@@ -46,6 +79,18 @@ const ShippingOptionButton = ({
           </p>
         )}
       </button>
+      {openPopover && (
+        <DeliveryPopover
+          onClick={handlePopoverClick}
+          handleOutSideClick={handleOutSideClick}
+          onChange={onChange ?? (() => {})}
+          onSubmit={onSubmit ?? (() => {})}
+          isLoading={loading}
+          inputErrorMessage={inputErrorMessage}
+          zipCode={zipCode}
+          variant={overlayType}
+        />
+      )}
     </div>
   )
 }