+++ CHANGELOG.md
@@ -6,9 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+
+- Popover to open the location drawer;
+- Popover with input to add postal code.
 
 ## [0.4.2] - 2024-11-06
-
 
 ### Fixed
 
@@ -19,7 +22,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Drawer opens behind the sidebar
-
 
 ## [0.4.0] - 2024-10-04